requirements.txt
python-telegram-bot==21.7
reportlab==4.0.4
google-cloud-speech==2.21.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
Flask==2.3.3
gunicorn==21.2.0